+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 - WebConfigPropertyCollection
   - Allowed different property collection key types to be added beyond the default.
+  - Allowed control over single item property collection key types, including examples - fixes ([issue #379](https://github.com/dsccommunity/WebAdministrationDsc/issues/379)), ([issue #631](https://github.com/dsccommunity/WebAdministrationDsc/issues/631)).
 
 ### Changed
 
@@ -23,6 +24,10 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 - WebAdministrationDsc
   - Fixed CertificateStoreName default value from `MY` to `My` ([issue #642](https://github.com/dsccommunity/WebAdministrationDsc/issues/642))
+
+### Removed
+
+- Removed outdated resources documentation from README.md.
 
 ## [4.2.0] - 2024-08-26